requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
python-dotenv
google-generativeai
chromadb
sentence-transformers
redis
black